MET-4233 Moved the url verification into controller
