+++ metis-dereference/metis-dereference-import/src/main/java/eu/europeana/metis/dereference/vocimport/VocabularyCollectionMavenRule.java
@@ -82,16 +82,6 @@ public class VocabularyCollectionMavenRule implements EnforcerRule {
   }
 
   /**
-   * Constructor. Used to inject the factory
-   *
-   * @param vocabularyCollectionImporterFactory The vocabulary collection importer factory
-   */
-  @Autowired
-  public VocabularyCollectionMavenRule(VocabularyCollectionImporterFactory vocabularyCollectionImporterFactory){
-    this.vocabularyCollectionImporterFactory = vocabularyCollectionImporterFactory;
-  }
-
-  /**
    * Constructor.
    *
    * @param lenientOnLackOfExamples Whether the the rule is lenient on vocabulary mappings without
