+++ metis-dereference/metis-dereference-service/src/main/java/eu/europeana/metis/dereference/service/MongoDereferencingManagementService.java
@@ -23,7 +23,6 @@ public class MongoDereferencingManagementService implements DereferencingManagem
 
   private final VocabularyDao vocabularyDao;
   private final ProcessedEntityDao processedEntityDao;
-  private final VocabularyCollectionImporterFactory vocabularyCollectionImporterFactory;
 
   /**
    * Constructor.
@@ -32,10 +31,9 @@ public class MongoDereferencingManagementService implements DereferencingManagem
    */
   @Autowired
   public MongoDereferencingManagementService(VocabularyDao vocabularyDao,
-          ProcessedEntityDao processedEntityDao, VocabularyCollectionImporterFactory vocabularyCollectionImporterFactory) {
+          ProcessedEntityDao processedEntityDao) {
     this.vocabularyDao = vocabularyDao;
     this.processedEntityDao = processedEntityDao;
-    this.vocabularyCollectionImporterFactory = vocabularyCollectionImporterFactory;
   }
 
   @Override
@@ -54,7 +52,7 @@ public class MongoDereferencingManagementService implements DereferencingManagem
     try {
     // Import and validate the vocabularies
     final List<Vocabulary> vocabularies = new ArrayList<>();
-    final VocabularyCollectionImporter importer = vocabularyCollectionImporterFactory
+    final VocabularyCollectionImporter importer = new VocabularyCollectionImporterFactory()
             .createImporter(directoryUrl);
     final VocabularyCollectionValidator validator = new VocabularyCollectionValidatorImpl(importer,
             true, true, true);
